Move Form inputs and basic structure into new component; update CreateOrphanage form to only handle specific submit action
